+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/query/QueryableFinatraCompositeWindowStore.scala
@@ -1,140 +1,128 @@
 package com.twitter.finatra.streams.query
 
+import com.twitter.finatra.streams.converters.time._
 import com.twitter.finatra.streams.queryable.thrift.domain.ServiceShardId
 import com.twitter.finatra.streams.queryable.thrift.partitioning.{
   KafkaPartitioner,
   StaticServiceShardPartitioner
 }
+import com.twitter.finatra.streams.stores.internal.FinatraStoresGlobalManager
 import com.twitter.finatra.streams.transformer.FinatraTransformer.{DateTimeMillis, WindowStartTime}
 import com.twitter.finatra.streams.transformer.domain.{CompositeKey, Time, TimeWindowed}
 import com.twitter.inject.Logging
 import com.twitter.util.Duration
 import org.apache.kafka.common.serialization.{Serde, Serializer}
-import org.apache.kafka.streams.KafkaStreams
-import org.apache.kafka.streams.errors.InvalidStateStoreException
-import org.apache.kafka.streams.state.{KeyValueIterator, QueryableStoreTypes, ReadOnlyKeyValueStore}
-import org.joda.time.{DateTime, DateTimeUtils}
+import org.joda.time.DateTimeUtils
 import scala.collection.JavaConverters._
-import scala.collection.mutable
 
 //TODO: DRY with other queryable finatra stores
-class QueryableFinatraCompositeWindowStore[K, PK, SK, V](
-  kafkaStreams: => KafkaStreams,
+class QueryableFinatraCompositeWindowStore[PK, SK, V](
   storeName: String,
+  windowSize: Duration,
   primaryKeySerde: Serde[PK],
   numShards: Int,
   numQueryablePartitions: Int,
   currentShardId: Int)
     extends Logging {
+
+  // The number of windows to query before and/or after specified start and end times
+  private val defaultWindowMultiplier = 3
+
   private val primaryKeySerializer = primaryKeySerde.serializer()
 
   private val currentServiceShardId = ServiceShardId(currentShardId)
+
+  private val windowSizeMillis = windowSize.inMillis
 
   private val partitioner = new KafkaPartitioner(
     StaticServiceShardPartitioner(numShards = numShards),
     numPartitions = numQueryablePartitions
   )
 
-  private val queryableStoreType = {
-    QueryableStoreTypes.keyValueStore[TimeWindowed[CompositeKey[PK, SK]], V]
-  }
+  /* Public */
 
-  //TODO: DRY with other queryable stores
+  /**
+   * Get a range of composite keys and return them combined in a map. If the primary key for the
+   * composite keys is non-local to this Kafka Streams instance, return an exception indicating
+   * which instance is hosting this primary key
+   *
+   * @param primaryKey The primary key for the data being queried (e.g. a UserId)
+   * @param startCompositeKey The starting composite key being queried (e.g. UserId-ClickType)
+   * @param endCompositeKey The ending composite key being queried (e.g. UserId-ClickType)
+   * @param allowStaleReads Allow stale reads when querying a caching key value store. If set to false,
+   *                        each query will trigger a flush of the cache.
+   * @param startTime The start time of the windows being queried
+   * @param endTime The end time of the windows being queried
+   * @return A time windowed map of composite keys to their values
+   */
   def get(
     primaryKey: PK,
-    windowSize: Duration,
-    startTime: Option[Long],
-    endTime: Option[Long],
-    queryLocalWindow: (PK, DateTime,
-      DateTime) => KeyValueIterator[TimeWindowed[CompositeKey[PK, SK]], V]
-  ): Map[WindowStartTime, Map[SK, V]] = {
+    startCompositeKey: CompositeKey[PK, SK],
+    endCompositeKey: CompositeKey[PK, SK],
+    allowStaleReads: Boolean,
+    startTime: Option[DateTimeMillis] = None,
+    endTime: Option[DateTimeMillis] = None
+  ): Map[WindowStartTime, scala.collection.Map[SK, V]] = {
     throwIfNonLocalKey(primaryKey, primaryKeySerializer)
-
-    val windowSizeMillis = windowSize.inMillis
 
     val (startWindowRange, endWindowRange) = startAndEndRange(
       startTime = startTime,
       endTime = endTime,
-      windowSizeMillis = windowSizeMillis
-    )
+      windowSizeMillis = windowSizeMillis)
 
-    val windowedCountsMapBuilder = Map.newBuilder[WindowStartTime, Map[SK, V]]
-    val numWindows = (endWindowRange - startWindowRange) / windowSizeMillis
-    windowedCountsMapBuilder.sizeHint(numWindows.toInt)
+    val resultMap = new java.util.TreeMap[Long, scala.collection.mutable.Map[SK, V]]().asScala
 
-    var currentWindowStart = startWindowRange
-    var currentWindowEnd = currentWindowStart + windowSizeMillis
-    while (currentWindowStart <= endWindowRange) {
-      callQueryLocal(
-        primaryKey,
-        queryLocalWindow,
-        currentWindowStart,
-        currentWindowEnd,
-        windowedCountsMapBuilder
+    var windowStartTime = startWindowRange
+    while (windowStartTime <= endWindowRange) {
+      queryWindow(startCompositeKey, endCompositeKey, windowStartTime, allowStaleReads, resultMap)
+      windowStartTime = windowStartTime + windowSizeMillis
+    }
+
+    resultMap.toMap
+  }
+
+  /* Private */
+
+  private def queryWindow(
+    startCompositeKey: CompositeKey[PK, SK],
+    endCompositeKey: CompositeKey[PK, SK],
+    windowStartTime: DateTimeMillis,
+    allowStaleReads: Boolean,
+    resultMap: scala.collection.mutable.Map[Long, scala.collection.mutable.Map[SK, V]]
+  ): Unit = {
+    trace(s"QueryWindow $startCompositeKey to $endCompositeKey ${windowStartTime.iso8601}")
+
+    //TODO: Use store.taskId to find exact store where the key is assigned
+    for (store <- FinatraStoresGlobalManager.getWindowedCompositeStores[PK, SK, V](storeName)) {
+      val iterator = store.range(
+        TimeWindowed.forSize(startMs = windowStartTime, windowSizeMillis, startCompositeKey),
+        TimeWindowed.forSize(startMs = windowStartTime, windowSizeMillis, endCompositeKey),
+        allowStaleReads = allowStaleReads
       )
 
-      currentWindowStart = currentWindowStart + windowSizeMillis
-      currentWindowEnd = currentWindowEnd + windowSizeMillis
-    }
-
-    windowedCountsMapBuilder.result()
-  }
-
-  def store: ReadOnlyKeyValueStore[TimeWindowed[CompositeKey[PK, SK]], V] = {
-    if (kafkaStreams == null) {
-      throw new InvalidStateStoreException(s"Store $storeName not started")
-    }
-
-    kafkaStreams.store(storeName, queryableStoreType)
-  }
-
-  private def callQueryLocal(
-    primaryKey: PK,
-    queryLocalWindow: (PK, DateTime,
-      DateTime) => KeyValueIterator[TimeWindowed[CompositeKey[PK, SK]], V],
-    currentWindowStart: DateTimeMillis,
-    currentWindowEnd: DateTimeMillis,
-    windowedCountsMapBuilder: mutable.Builder[
-      (WindowStartTime, Map[SK, V]),
-      Map[
-        WindowStartTime,
-        Map[SK, V]
-      ]]
-  ) = {
-    val localEntriesIterator =
-      queryLocalWindow(primaryKey, new DateTime(currentWindowStart), new DateTime(currentWindowEnd))
-
-    try {
-      val localEntries = localEntriesIterator.asScala
-      val attributeMapBuilder = Map.newBuilder[SK, V]
-      attributeMapBuilder.sizeHint(8)
-      for (entry <- localEntries) {
-        attributeMapBuilder += (entry.key.value.secondary -> entry.value)
+      while (iterator.hasNext) {
+        val entry = iterator.next()
+        trace(s"$store\t$entry")
+        val innerMap =
+          resultMap.getOrElseUpdate(entry.key.startMs, scala.collection.mutable.Map[SK, V]())
+        innerMap += (entry.key.value.secondary -> entry.value)
       }
-
-      val attributeMap = attributeMapBuilder.result()
-      if (attributeMap.nonEmpty) {
-        windowedCountsMapBuilder += (currentWindowStart -> attributeMap)
-      }
-    } finally {
-      localEntriesIterator.close()
     }
   }
 
-  //TODO: Reuse in other queryable stores
   private def startAndEndRange(
     startTime: Option[DateTimeMillis],
     endTime: Option[DateTimeMillis],
     windowSizeMillis: DateTimeMillis
-  ) = {
-    val endWindowRange = endTime.getOrElse(
+  ): (DateTimeMillis, DateTimeMillis) = {
+    val endWindowRange = endTime.getOrElse {
       TimeWindowed.windowStart(
         messageTime = Time(DateTimeUtils.currentTimeMillis),
-        sizeMs = windowSizeMillis
-      ) + windowSizeMillis
-    )
+        sizeMs = windowSizeMillis) + defaultWindowMultiplier * windowSizeMillis
+    }
 
-    val startWindowRange = startTime.getOrElse(endWindowRange - (3 * windowSizeMillis))
+    val startWindowRange =
+      startTime.getOrElse(endWindowRange - (defaultWindowMultiplier * windowSizeMillis))
 
     (startWindowRange, endWindowRange)
   }