+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/streams/transformer/domain/FixedTimeWindowedSerde.scala
@@ -26,21 +26,23 @@ class FixedTimeWindowedSerde[K](val inner: Serde[K], windowSize: Duration)
   private val windowSizeMillis = windowSize.inMillis
   assert(windowSizeMillis > 10, "The minimum window size currently supported is 10ms")
 
+  /* Public */
+
   final override def deserialize(bytes: Array[Byte]): TimeWindowed[K] = {
     val keyBytesSize = bytes.length - WindowStartTimeSizeBytes
     val keyBytes = new Array[Byte](keyBytesSize)
 
     val bb = ByteBuffer.wrap(bytes)
     val startMs = bb.getLong()
-    val endMs = startMs + windowSizeMillis
     bb.get(keyBytes)
+    val endMs = startMs + windowSizeMillis
 
     TimeWindowed(startMs = startMs, endMs = endMs, innerDeserializer.deserialize(topic, keyBytes))
   }
 
   final override def serialize(timeWindowedKey: TimeWindowed[K]): Array[Byte] = {
     assert(
-      timeWindowedKey.endMs == Long.MaxValue || timeWindowedKey.startMs + windowSizeMillis == timeWindowedKey.endMs,
+      timeWindowedKey.startMs + windowSizeMillis == timeWindowedKey.endMs,
       s"TimeWindowed element being serialized has end time which is not consistent with the FixedTimeWindowedSerde window size of $windowSize. ${timeWindowedKey.startMs + windowSizeMillis} != ${timeWindowedKey.endMs}"
     )
 
@@ -48,21 +50,8 @@ class FixedTimeWindowedSerde[K](val inner: Serde[K], windowSize: Duration)
     val windowAndKeyBytesSize = new Array[Byte](WindowStartTimeSizeBytes + keyBytes.length)
 
     val bb = ByteBuffer.wrap(windowAndKeyBytesSize)
-    bb.putLong(startMs(timeWindowedKey))
+    bb.putLong(timeWindowedKey.startMs)
     bb.put(keyBytes)
     bb.array()
-  }
-
-  /*
-   * We need to special case fixed fixed windows when the endMs is Long.MaxValue which signals all elements in this window
-   * We special case because this fixed window serde doesn't serialize endMs to save bytes
-   * Note: We use this value in keyValueStore.range which treats the to and from times as both inclusive (which differs from TimeWindowed which is documented as from being exclusive...)
-   */
-  private def startMs(timeWindowedKey: TimeWindowed[K]) = {
-    if (timeWindowedKey.endMs == Long.MaxValue) {
-      timeWindowedKey.startMs + 1
-    } else {
-      timeWindowedKey.startMs
-    }
   }
 }