+++ kafka-streams/kafka-streams/src/test/scala/com/twitter/unittests/integration/compositesum/UserClicks.scala
@@ -1,0 +1,5 @@
+package com.twitter.unittests.integration.compositesum
+
+import com.twitter.unittests.integration.compositesum.UserClicksTypes.UserId
+
+case class UserClicks(userId: UserId, clickType: Int)